layers: Move device memory CoreCheck to own file
