+++ BUILD.gn
@@ -129,6 +129,7 @@ core_validation_sources = [
   "layers/descriptor_validation.cpp",
   "layers/device_memory_state.cpp",
   "layers/device_memory_state.h",
+  "layers/device_memory_validation.cpp",
   "layers/device_state.h",
   "layers/drawdispatch.cpp",
   "layers/generated/command_validation.cpp",
